Update README.md (#530)
* Update README.md

The default behaviour of where the convert.py saved files was wrong. It also was inconsistent with how the later script test.py is trying to use them (and assuming naming convention). 

I don't actually see a quick way to automate this since--as written--the  target directory is set directly by an argument. It would probably be best to rewrite it so that the argument is used as an override variable, but the default behaviour is to construct a file path based on set and unset arugments. This also is complex because "defaults" are assumed in the naming convention as well.

* Update README.md

Created an actual script that'll run and do this correctly.

* Update README.md

Typo fix: mlx-models should have been mlx_models. This conforms with standard later in the mlx-examples/whisper code.

* Update README.md

Removed the larger script and changed it back to the simpler script as before.

* nits in readme

---------

Co-authored-by: Awni Hannun <awni@apple.com>