Update utils.py
To fit with Llama-3-8B-Instruct Model, which have 2 types of EOS tokens ("<|end_of_text|>" and "<|eot_id|>"), I add an optional parameter (eos_token_ids) of generate function to receive an array of EOS tokens for Llama-3-8B-Instruct Model and changed the for-loop break check as "if token in eos_token_ids:". If the function does not receive any eos_token_ids as an input, it is by default using "[tokenizer.eos_token_id]".